backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Basketball statistics dashboard backend"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Project Maintainer", email = "maintainer@example.com"},
]
license = {text = "MIT"}

dependencies = [
  "fastapi",
  "uvicorn",
  "duckdb",
  "pydantic",
  "python-multipart",
  "pandas",
  "numpy",
  "slowapi",
  "python-json-logger>=3.0.0",
  "strawberry-graphql[fastapi]",
  "nba-api>=1.11.3",
  "basketball-reference-web-scraper>=4.15.4",
  "pydantic-settings>=2.12.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/basketball-reference-clone"
Repository = "https://github.com/yourusername/basketball-reference-clone"

[dependency-groups]
dev = [
  "mypy",
  "ruff",
  "pytest",
  "pytest-cov",
  "types-python-dateutil",
  "pandas-stubs",
  "make>=0.1.6.post2",
  "httpx>=0.28.1",
]

testing = [
  "pytest",
  "pytest-cov",
  "httpx>=0.28.1",
  "pytest-asyncio",
]

linting = [
  "ruff",
  "mypy",
  "pandas-stubs",
  "types-python-dateutil",
]

docs = [
  "mkdocs",
  "mkdocs-material",
]

[project.optional-dependencies]
dev = [
  "mypy",
  "ruff",
  "pytest",
  "types-python-dateutil",
  "pandas-stubs",
  "httpx>=0.28.1",
  "pytest-asyncio",
]

[tool.ruff]
# Set the maximum line length to 100
line-length = 100
# Target Python 3.10+
target-version = "py310"

# Include Python files
include = ["*.py", "*.pyi"]

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
    "node_modules",
]

[tool.ruff.lint]
# Enable recommended rule sets
select = [
  "E",
  "W",
  "F",
  "N",
  "UP",
  "B",
  "C4",
  "SIM",
  "RUF",
  "D",
  "ANN",
  "ASYNC",
  "S",
  "BLE",
  "A",
  "COM",
  "C90",
  "DTZ",
  "EM",
  "FA",
  "ISC",
  "ICN"
]
ignore = [
  "E501",
  "B008",
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D107",
  "D203",
  "D213",
  "ANN003",
  "ANN002",
  "S101",
  "S105",
  "S106",
  "S107",
  "S311",
  "PLR0913",
  "PLR0911",
  "PLR0912",
  "PLR0915",
  "PLW2901",
  "PLW0603",
  "BLE001",
  "FBT001",
  "FBT002",
  "DTZ005",
  "EM101",
  "EM102",
  "G004",
  "INP001",
  "T201",
  "T203",
  "TD002",
  "TD003",
  "FIX002",
  "ERA001",
  "PGH003",
  "PGH004"
]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = ["scripts/research", "scripts/db_inspection"]
follow_imports = "silent"

[[tool.mypy.overrides]]
module = [
    "slowapi.*",
    "pythonjsonlogger.*",
    "nba_api.*"
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "transformers.*"
follow_imports = "skip"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "app/scripts/*",
    "app/api/graphql/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
